address review points of blake
